test(web): retire the flaky daemon markdown typing test for a jsdom forwarding test
DaemonDocumentPage.markdown.browser.test.tsx was the one remaining
known-flaky page-level browser test (~16s, real-CodeMirror focus + typing).
Its actual subject — a body edit travels the sync session's ordinary
set-body command path to backend.pushLocalUpdate on the hydrated doc
lineage — never needed the keyboard: the new jsdom test drives
MarkdownEditor's controlled onChange through a textarea stub and keeps the
same replay-over-snapshot assertion on the `body` text container.

Real-CodeMirror input coverage stays where it already lives:
loro-binding.browser.test.tsx pins typing-to-Loro independently, and the
editor-selection assertions stay in DaemonDocumentPage.markdown.test.tsx.

Mutation-checked: making the session's set-body handler a no-op fails the
new test; restored after.
